Create a new module token_operation
