+++ mtk-automation/src/contract_actions/compensate.rs
@@ -2,13 +2,11 @@
 //!
 //! Module for compensates control in the contract
 use soroban_sdk::{Env, Symbol, Map, AccountId};
-use soroban_auth::{Identifier, Signature};
+use soroban_auth::Signature;
 use crate::contract_actions::validation;
 use crate::contract_actions::datakey::DataKey;
 use crate::contract_actions::identifier_wrapper;
-use crate::contract_actions::token_contract;
-use crate::contract_actions::admin;
-use crate::contract_actions::token;
+use crate::contract_actions::token_operation;
 
 pub fn set_compensations(env: &Env, compensate_types: &Map<Symbol, u32>) {
     env.storage().set(DataKey::Compensate, compensate_types);
@@ -24,7 +22,7 @@ pub fn compensate_member(env: &Env, approval_sign: &Signature, to: &AccountId, c
     }
 
     let compensate_value = get_compensate_by_type(&env, &compensate_types);
-    transfer(&env, &approval_sign, &identifier_wrapper::get_account_identifier(to.clone()), &compensate_value);
+    token_operation::transfer(&env, &approval_sign, &identifier_wrapper::get_account_identifier(to.clone()), &compensate_value);
 }
 
 fn get_compensate_by_type(env: &Env, r_type: &Symbol) -> i128 {
@@ -32,16 +30,6 @@ fn get_compensate_by_type(env: &Env, r_type: &Symbol) -> i128 {
     let compensate: Map<Symbol, i128> = env.storage().get(key).unwrap().unwrap();
 
     compensate.get(r_type.clone()).unwrap().unwrap()
-}
-
-fn transfer(env: &Env, approval_sign: &Signature, to: &Identifier, amount: &i128) {
-    let tc_id = token_contract::get_token_contract_id(&env);
-    let client = token::Client::new(&env, tc_id);
-
-    let admin_id = admin::get_admin_id(&env);
-    let nonce = client.nonce(&admin_id);
-
-    client.xfer(&approval_sign, &nonce, &to, &amount);
 }
 
 fn get_compensates(env: &Env) -> Map<Symbol, u32> {
