+++ mtk-automation/src/contract_actions/mod.rs
@@ -6,6 +6,7 @@ pub mod admin;
 pub mod identifier_wrapper;
 pub mod compensate;
 pub mod validation;
+pub mod token_operation;
 pub mod token {
   soroban_sdk::contractimport!(file = "./soroban_token_spec.wasm");
 }
